bridges/requirements.txt
# MQTT/NATS Bridge Requirements
asyncio-nats-client>=0.14.0
asyncio-mqtt>=0.16.0
paho-mqtt>=1.6.0
aiofiles>=23.0.0
pydantic>=2.0.0
python-json-logger>=2.0.0
click>=8.1.0
colorlog>=6.7.0
prometheus-client>=0.19.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
pylint>=2.17.0
mypy>=1.5.0
isort>=5.12.0